+++ Scan_File_Processor_Importer/src_main/org/yeastrc/spectral_storage/scan_file_processor/input_scan_file/dto/ScanPeak.java
@@ -2,10 +2,10 @@ package org.yeastrc.spectral_storage.scan_file_processor.input_scan_file.dto;
 
 public class ScanPeak {
 
-    private float mz;
+    private double mz;
     private float intensity;
     
-	public void setMz(float mz) {
+	public void setMz(double mz) {
 		this.mz = mz;
 	}
 	public void setIntensity(float intensity) {
@@ -13,7 +13,7 @@ public class ScanPeak {
 	}
 
 	
-	public float getMz() {
+	public double getMz() {
 		return mz;
 	}
 	public float getIntensity() {
