+++ Scan_File_Processor_Importer/src_mzML_Reader/org/systemsbiology/jrap/stax/ScanAndHeaderParser.java
@@ -178,7 +178,7 @@ public class ScanAndHeaderParser{
 			    elementName = xmlSR.getLocalName();
 			    if(elementName.equals("precursorMz"))
 				{
-				    tmpScanHeader.setPrecursorMz(Float.parseFloat(precursorBuffer.toString()));
+				    tmpScanHeader.setPrecursorMz(Double.parseDouble(precursorBuffer.toString()));
                                 
 				    precursorBuffer = null; // make available for garbage collection
