+++ WebService_Web_App/src_shared_server_client/org/yeastrc/spectral_storage/shared_server_client/webservice_request_response/sub_parts/SingleScan_SubResponse.java
@@ -31,7 +31,7 @@ public class SingleScan_SubResponse {
 	@XmlAttribute // attribute name is property name
 	private Byte precursorCharge;
 	@XmlAttribute // attribute name is property name
-	private Float precursor_M_Over_Z;
+	private Double precursor_M_Over_Z;
 	
 	// Peaks
 	@XmlElementWrapper(name="peaks")
@@ -82,10 +82,10 @@ public class SingleScan_SubResponse {
 	public void setPrecursorCharge(Byte precursorCharge) {
 		this.precursorCharge = precursorCharge;
 	}
-	public Float getPrecursor_M_Over_Z() {
+	public Double getPrecursor_M_Over_Z() {
 		return precursor_M_Over_Z;
 	}
-	public void setPrecursor_M_Over_Z(Float precursor_M_Over_Z) {
+	public void setPrecursor_M_Over_Z(Double precursor_M_Over_Z) {
 		this.precursor_M_Over_Z = precursor_M_Over_Z;
 	}
