+++ WebService_Web_App/src_spectral_file/org/yeastrc/spectral_storage/spectral_file_common/spectral_file/storage_files_on_disk/version_003/data_file/on_disk_dto/SpectralFile_SingleScanPeak_V_003.java
@@ -8,7 +8,7 @@ package org.yeastrc.spectral_storage.spectral_file_common.spectral_file.storage_
  */
 public class SpectralFile_SingleScanPeak_V_003 {
 
-	private float m_over_Z;
+	private double m_over_Z;
 	private float intensity;
 	
 }
